requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
Flask-Mail==0.10.0
Pillow==10.4.0
bleach==6.1.0
